+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- BREAKING: Change accounts passed to `cache_oracle`
+
 ## [0.5.0] - 2022-04-25
 
 - Add `taker_rate` and `maker_rate` functions